변유송#12.10/16:44
OCR 폴더 이동
- EMBRASSER views.py 수정
- EMBRASSER urls.py 수정
- OCR views.py 수정
- OCR urls.py 수정
- index.html 수정
- member_ocr.html 수정
