+++ EMBRASSER/template/index.html
@@ -8,7 +8,7 @@
 
         <h1>
             {{menutitle}}
-            <a href="coocr_upload">OCR</a>
+            <a href="ocr/coocr_upload">OCR</a>
         </h1>
 
     </body>
