+++ EMBRASSER/template/member_ocr.html
@@ -8,11 +8,11 @@
 
         <h1>
             {{menutitle}}
-            <a href="coocr_upload">OCR</a>
+            <a href="ocr/coocr_upload">OCR</a>
         </h1>
 
         <form
-            action="/coocr_upload"
+            action="/ocr/coocr_upload"
             method="post"
             id="Coauthform"
             enctype="multipart/form-data">
